Divide organisation tabs in single components
