+++ frontend/src/app/pages/organisations/tabs/settings/organisation-tab-settings.component.html
@@ -1,0 +1,20 @@
+<div class="card border-danger mt-2">
+    <div class="card-body d-flex align-items-center">
+        <div class="flex-grow-1">
+            <h5 class="card-title text-danger">Organisation verlassen</h5>
+            Wenn Sie die Organisation verlassen, haben Sie keinen Zugriff mehr
+            auf die dort gespeicherten Daten.
+        </div>
+        <button class="btn btn-danger" (click)="leave()">Verlassen</button>
+    </div>
+</div>
+<div class="card border-danger mt-2">
+    <div class="card-body d-flex align-items-center">
+        <div class="flex-grow-1">
+            <h5 class="card-title text-danger">Organisation löschen</h5>
+            Wenn Sie die Organisation löschen, werden alle dort gespeicherten
+            Daten unwiderruflich gelöscht.
+        </div>
+        <button class="btn btn-danger" (click)="delete()">Löschen</button>
+    </div>
+</div>
